NHL Playoff Carousel

Returns playoff series carousel data for a given season.

Usage

nhl_playoff_carousel(season = NULL)

Arguments

season

Integer 4-digit year (e.g., 2024 for the 2024-25 season). If NULL, uses most recent season.

Value

Returns a list with playoff series carousel data.

Examples

# \donttest{
  try(nhl_playoff_carousel(season = 2024))
#> $seasonId
#> [1] 20242025
#> 
#> $currentRound
#> [1] 4
#> 
#> $rounds
#>   roundNumber        roundLabel roundAbbrev
#> 1           1         1st-round          R1
#> 2           2         2nd-round          R2
#> 3           3 conference-finals          CF
#> 4           4 stanley-cup-final         SCF
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            series
#> 1 A, B, C, D, E, F, G, H, 1, 1, 1, 1, 1, 1, 1, 1, 1st-round, 1st-round, 1st-round, 1st-round, 1st-round, 1st-round, 1st-round, 1st-round, /schedule/playoff-series/2025/series-a/senators-vs-mapleleafs, /schedule/playoff-series/2025/series-b/panthers-vs-lightning, /schedule/playoff-series/2025/series-c/canadiens-vs-capitals, /schedule/playoff-series/2025/series-d/devils-vs-hurricanes, /schedule/playoff-series/2025/series-e/blues-vs-jets, /schedule/playoff-series/2025/series-f/avalanche-vs-stars, /schedule/playoff-series/2025/series-g/wild-vs-goldenknights, /schedule/playoff-series/2025/series-h/oilers-vs-kings, 4, 4, 4, 4, 4, 4, 4, 4, 10, 13, 15, 12, 52, 25, 54, 22, 9, 14, 8, 1, 19, 21, 30, 26, 9, 13, 8, 1, 19, 21, 30, 22, OTT, FLA, MTL, NJD, STL, COL, MIN, EDM, 2, 4, 1, 1, 3, 3, 2, 4, https://assets.nhle.com/logos/nhl/svg/OTT_light.svg, https://assets.nhle.com/logos/nhl/svg/FLA_light.svg, https://assets.nhle.com/logos/nhl/svg/MTL_light.svg, https://assets.nhle.com/logos/nhl/svg/NJD_light.svg, https://assets.nhle.com/logos/nhl/svg/STL_20082009-20242025_light.svg, https://assets.nhle.com/logos/nhl/svg/COL_light.svg, https://assets.nhle.com/logos/nhl/svg/MIN_light.svg, https://assets.nhle.com/logos/nhl/svg/EDM_light.svg, https://assets.nhle.com/logos/nhl/svg/OTT_dark.svg, https://assets.nhle.com/logos/nhl/svg/FLA_dark.svg, https://assets.nhle.com/logos/nhl/svg/MTL_dark.svg, https://assets.nhle.com/logos/nhl/svg/NJD_dark.svg, https://assets.nhle.com/logos/nhl/svg/STL_20082009-20242025_dark.svg, https://assets.nhle.com/logos/nhl/svg/COL_dark.svg, https://assets.nhle.com/logos/nhl/svg/MIN_dark.svg, https://assets.nhle.com/logos/nhl/svg/EDM_dark.svg, 10, 14, 15, 12, 52, 25, 54, 26, TOR, TBL, WSH, CAR, WPG, DAL, VGK, LAK, 4, 1, 4, 4, 4, 4, 4, 2, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, https://assets.nhle.com/logos/nhl/svg/TBL_light.svg, https://assets.nhle.com/logos/nhl/svg/WSH_secondary_light.svg, https://assets.nhle.com/logos/nhl/svg/CAR_light.svg, https://assets.nhle.com/logos/nhl/svg/WPG_light.svg, https://assets.nhle.com/logos/nhl/svg/DAL_light.svg, https://assets.nhle.com/logos/nhl/svg/VGK_light.svg, https://assets.nhle.com/logos/nhl/svg/LAK_light.svg, https://assets.nhle.com/logos/nhl/svg/TOR_dark.svg, https://assets.nhle.com/logos/nhl/svg/TBL_dark.svg, https://assets.nhle.com/logos/nhl/svg/WSH_secondary_dark.svg, https://assets.nhle.com/logos/nhl/svg/CAR_dark.svg, https://assets.nhle.com/logos/nhl/svg/WPG_dark.svg, https://assets.nhle.com/logos/nhl/svg/DAL_dark.svg, https://assets.nhle.com/logos/nhl/svg/VGK_dark.svg, https://assets.nhle.com/logos/nhl/svg/LAK_dark.svg
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  I, J, K, L, 2, 2, 2, 2, 2nd-round, 2nd-round, 2nd-round, 2nd-round, /schedule/playoff-series/2025/series-i/panthers-vs-mapleleafs, /schedule/playoff-series/2025/series-j/hurricanes-vs-capitals, /schedule/playoff-series/2025/series-k/stars-vs-jets, /schedule/playoff-series/2025/series-l/oilers-vs-goldenknights, 4, 4, 4, 4, 13, 12, 25, 22, 10, 15, 52, 54, 13, 12, 25, 22, FLA, CAR, DAL, EDM, 4, 4, 4, 4, https://assets.nhle.com/logos/nhl/svg/FLA_light.svg, https://assets.nhle.com/logos/nhl/svg/CAR_light.svg, https://assets.nhle.com/logos/nhl/svg/DAL_light.svg, https://assets.nhle.com/logos/nhl/svg/EDM_light.svg, https://assets.nhle.com/logos/nhl/svg/FLA_dark.svg, https://assets.nhle.com/logos/nhl/svg/CAR_dark.svg, https://assets.nhle.com/logos/nhl/svg/DAL_dark.svg, https://assets.nhle.com/logos/nhl/svg/EDM_dark.svg, 10, 15, 52, 54, TOR, WSH, WPG, VGK, 3, 1, 2, 1, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, https://assets.nhle.com/logos/nhl/svg/WSH_secondary_light.svg, https://assets.nhle.com/logos/nhl/svg/WPG_light.svg, https://assets.nhle.com/logos/nhl/svg/VGK_light.svg, https://assets.nhle.com/logos/nhl/svg/TOR_dark.svg, https://assets.nhle.com/logos/nhl/svg/WSH_secondary_dark.svg, https://assets.nhle.com/logos/nhl/svg/WPG_dark.svg, https://assets.nhle.com/logos/nhl/svg/VGK_dark.svg
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               M, N, 3, 3, conference-finals, conference-finals, /schedule/playoff-series/2025/series-m/panthers-vs-hurricanes, /schedule/playoff-series/2025/series-n/oilers-vs-stars, 4, 4, 13, 22, 12, 25, 13, 22, FLA, EDM, 4, 4, https://assets.nhle.com/logos/nhl/svg/FLA_light.svg, https://assets.nhle.com/logos/nhl/svg/EDM_light.svg, https://assets.nhle.com/logos/nhl/svg/FLA_dark.svg, https://assets.nhle.com/logos/nhl/svg/EDM_dark.svg, 12, 25, CAR, DAL, 1, 1, https://assets.nhle.com/logos/nhl/svg/CAR_light.svg, https://assets.nhle.com/logos/nhl/svg/DAL_light.svg, https://assets.nhle.com/logos/nhl/svg/CAR_dark.svg, https://assets.nhle.com/logos/nhl/svg/DAL_dark.svg
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         O, 4, stanley-cup-final, /schedule/playoff-series/2025/series-o/panthers-vs-oilers, 4, 13, 22, 13, FLA, 4, https://assets.nhle.com/logos/nhl/svg/FLA_light.svg, https://assets.nhle.com/logos/nhl/svg/FLA_dark.svg, 22, EDM, 2, https://assets.nhle.com/logos/nhl/svg/EDM_light.svg, https://assets.nhle.com/logos/nhl/svg/EDM_dark.svg
#> 
# }