NHL Team Scoreboard

Returns current scoreboard information for a specific team, including upcoming and recent games.

Usage

nhl_team_scoreboard(team_abbr)

Arguments

team_abbr

Three-letter team abbreviation (e.g., "TOR", "BOS")

Value

A named list of data frames: gamesByDate.

gamesByDate

col_name	types	description

date	character	Date the games are scheduled for.

games	list	List of games scheduled on that date.

Examples

# \donttest{
  try(nhl_team_scoreboard(team_abbr = "TOR"))
#> $focusedDate
#> [1] "2026-10-08"
#> 
#> $focusedDateCount
#> [1] 10
#> 
#> $clubTimeZone
#> [1] "America/Toronto"
#> 
#> $clubUTCOffset
#> [1] "-04:00"
#> 
#> $clubScheduleLink
#> [1] "/mapleleafs/schedule"
#> 
#> $gamesByDate
#>          date
#> 1  2026-09-30
#> 2  2026-10-03
#> 3  2026-10-06
#> 4  2026-10-08
#> 5  2026-10-10
#> 6  2026-10-13
#> 7  2026-10-15
#> 8  2026-10-17
#> 9  2026-10-19
#> 10 2026-10-20
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        games
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            2026020008, 20262027, 2, 2026-09-30, /gamecenter/nyi-vs-tor/2026/09/30/2026020008, 2026-09-30T23:30:00Z, -04:00, -04:00, 548, 409, N, A, CA, US, Prime, MSGSN, 116, 411, https://assets.nhle.com/logos/broadcast/548-light.svg, NA, https://assets.nhle.com/logos/broadcast/548-dark.svg, NA, OFF, OK, 3, /video/nyi-at-tor-recap-6405995463112, /fr/video/nyi-vs-tor-30-09-2026-resume-6405995839112, Scotiabank Arena, 2, NYI, 1, https://assets.nhle.com/logos/nhl/svg/NYI_light.svg, New York Islanders, Islanders de New York, Islanders, New York, de New York, 10, TOR, 2, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 3, REG, 3
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        2026020023, 20262027, 2, 2026-10-03, /gamecenter/ott-vs-tor/2026/10/03/2026020023, 2026-10-03T23:00:00Z, -04:00, -04:00, 282, N, CA, SN, 21, en, 282, https://assets.nhle.com/logos/broadcast/282-light.svg, https://assets.nhle.com/logos/broadcast/282-dark.svg, OFF, OK, 3, /video/ott-at-tor-recap-6406196917112, /fr/video/ott-vs-tor-03-10-2026-resume-6406195574112, Scotiabank Arena, 9, OTT, 3, https://assets.nhle.com/logos/nhl/svg/OTT_light.svg, Ottawa Senators, Sénateurs d'Ottawa, Senators, Sénateurs, Ottawa, d'Ottawa, 10, TOR, 2, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 3, REG, 3
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            2026020044, 20262027, 2, 2026-10-06, /gamecenter/nsh-vs-tor/2026/10/06/2026020044, 2026-10-06T23:00:00Z, -04:00, -04:00, 293, 521, H, A, CA, US, TSN4, SCRIPPS, 137, 657, OFF, OK, 4, /video/nsh-at-tor-recap-6406392025112, /fr/video/nsh-vs-tor-06-10-2026-resume-6406393236112, Scotiabank Arena, 18, NSH, 4, https://assets.nhle.com/logos/nhl/svg/NSH_light.svg, Nashville Predators, Predators de Nashville, Predators, Nashville, de Nashville, 10, TOR, 5, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 4, OT, 3
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         2026020065, 20262027, 2, 2026-10-08, /gamecenter/tor-vs-vgk/2026/10/08/2026020065, 2026-10-09T02:00:00Z, -04:00, -07:00, 293, 521, A, H, CA, US, TSN4, SCRIPPS, 137, 657, FUT, OK, https://tix.axs.com/c8egKQAAAADI0LbuAgAAAADw%2fv%2f%2f%2fwD%2f%2f%2f%2f%2fEnZlZ2FzZ29sZGVua25pZ2h0cwD%2f%2f%2f%2f%2f%2f%2f%2f%2f%2fw%3d%3d, https://tix.axs.com/c8egKQAAAADI0LbuAgAAAADw%2fv%2f%2f%2fwD%2f%2f%2f%2f%2fEnZlZ2FzZ29sZGVua25pZ2h0cwD%2f%2f%2f%2f%2f%2f%2f%2f%2f%2fw%3d%3d, T-Mobile Arena, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 54, VGK, 3-1-0, https://assets.nhle.com/logos/nhl/svg/VGK_light.svg, Vegas Golden Knights, Golden Knights de Vegas, Golden Knights, Vegas, de Vegas
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                2026020080, 20262027, 2, 2026-10-10, /gamecenter/tor-vs-col/2026/10/10/2026020080, 2026-10-10T23:00:00Z, -04:00, -06:00, 282, 324, 2, N, N, H, CA, US, US, SN, NHLN, ALT, 21, 35, 385, en, NA, NA, 282, NA, NA, https://assets.nhle.com/logos/broadcast/282-light.svg, https://assets.nhle.com/logos/broadcast/324-light.svg, NA, https://assets.nhle.com/logos/broadcast/282-dark.svg, https://assets.nhle.com/logos/broadcast/324-dark.svg, NA, FUT, OK, https://avs.social/2627_SCOREBOARD_GM3, https://avs.social/2627_SCOREBOARD_GM3, Ball Arena, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 21, COL, 2-1-0, https://assets.nhle.com/logos/nhl/svg/COL_light.svg, Colorado Avalanche, Avalanche du Colorado, Avalanche, Colorado, du Colorado
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               2026020102, 20262027, 2, 2026-10-13, /gamecenter/tor-vs-uta/2026/10/13/2026020102, 2026-10-14T01:00:00Z, -04:00, -06:00, 293, 558, A, H, CA, US, TSN4, Utah16, 137, 381, FUT, OK, https://seatgeek.com/utah-mammoth-tickets/7-10-2026-salt-lake-city-utah-delta-center-1/nhl/18242273?wid=club-web-scoreboard, https://seatgeek.com/utah-mammoth-tickets/7-10-2026-salt-lake-city-utah-delta-center-1/nhl/18242273?wid=club-web-scoreboard, Delta Center, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 68, UTA, 3-1-0, https://assets.nhle.com/logos/nhl/svg/UTA_light.svg, Utah Mammoth, Mammoth de l'Utah, Mammoth, Utah, de l'Utah
#> 7                                                                                                                                                                                                                                                                                                                                                                       2026020108, 20262027, 2, 2026-10-15, /gamecenter/buf-vs-tor/2026/10/15/2026020108, 2026-10-15T23:00:00Z, -04:00, -04:00, 293, 28, H, A, CA, US, TSN4, MSG-B, 137, 392, FUT, OK, https://www.ticketmaster.ca/event/100064FECCD7F7C9?brand=torontomapleleafs&artistid=806033&wt.mc_id=NHL_TEAM_TOR_SCOREBOARD_GM5&utm_source=NHL.com&utm_medium=client_PM&utm_campaign=NHL_TEAM_TOR&utm_content=SCOREBOARD_GM5, https://www.ticketmaster.ca/event/100064FECCD7F7C9?brand=torontomapleleafs&artistid=806033&wt.mc_id=NHL_TEAM_TOR_SCOREBOARD_GM5&utm_source=NHL.com&utm_medium=client_PM&utm_campaign=NHL_TEAM_TOR&utm_content=SCOREBOARD_GM5, Scotiabank Arena, 7, BUF, 2-1-0, https://assets.nhle.com/logos/nhl/svg/BUF_light.svg, Buffalo Sabres, Sabres de Buffalo, Sabres, Buffalo, de Buffalo, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto
#> 8                                                                                                                                                                                                                      2026020125, 20262027, 2, 2026-10-17, /gamecenter/nyi-vs-tor/2026/10/17/2026020125, 2026-10-17T23:00:00Z, -04:00, -04:00, 282, 409, N, A, CA, US, SN, MSGSN, 21, 411, en, NA, 282, NA, https://assets.nhle.com/logos/broadcast/282-light.svg, NA, https://assets.nhle.com/logos/broadcast/282-dark.svg, NA, FUT, OK, https://www.ticketmaster.ca/event/100064FECCDCF7DC?brand=torontomapleleafs&artistid=806033&wt.mc_id=NHL_TEAM_TOR_SCOREBOARD_GM6&utm_source=NHL.com&utm_medium=client_PM&utm_campaign=NHL_TEAM_TOR&utm_content=SCOREBOARD_GM6, https://www.ticketmaster.ca/event/100064FECCDCF7DC?brand=torontomapleleafs&artistid=806033&wt.mc_id=NHL_TEAM_TOR_SCOREBOARD_GM6&utm_source=NHL.com&utm_medium=client_PM&utm_campaign=NHL_TEAM_TOR&utm_content=SCOREBOARD_GM6, Scotiabank Arena, 2, NYI, 1-2-0, https://assets.nhle.com/logos/nhl/svg/NYI_light.svg, New York Islanders, Islanders de New York, Islanders, New York, de New York, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto
#> 9                                                                                                        2026020139, 20262027, 2, 2026-10-19, /gamecenter/sjs-vs-tor/2026/10/19/2026020139, 2026-10-19T23:30:00Z, -04:00, -04:00, 549, 281, 314, N, N, A, CA, CA, US, SN+, TVAS, NBCSCA, 117, 120, 393, https://assets.nhle.com/logos/broadcast/549-light.svg, https://assets.nhle.com/logos/broadcast/281-light.svg, NA, https://assets.nhle.com/logos/broadcast/549-dark.svg, https://assets.nhle.com/logos/broadcast/281-dark.svg, NA, FUT, OK, https://www.ticketmaster.ca/event/100064FECCE1F7F0?brand=torontomapleleafs&artistid=806033&wt.mc_id=NHL_TEAM_TOR_SCOREBOARD_GM7&utm_source=NHL.com&utm_medium=client_PM&utm_campaign=NHL_TEAM_TOR&utm_content=SCOREBOARD_GM7, https://www.ticketmaster.ca/event/100064FECCE1F7F0?brand=torontomapleleafs&artistid=806033&wt.mc_id=NHL_TEAM_TOR_SCOREBOARD_GM7&utm_source=NHL.com&utm_medium=client_PM&utm_campaign=NHL_TEAM_TOR&utm_content=SCOREBOARD_GM7, Scotiabank Arena, 28, SJS, 2-1-0, https://assets.nhle.com/logos/nhl/svg/SJS_light.svg, San Jose Sharks, Sharks de San Jose, Sharks, San Jose, de San Jose, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto
#> 10 2026020145, 20262027, 2, 2026-10-20, /gamecenter/tor-vs-cbj/2026/10/20/2026020145, 2026-10-20T23:00:00Z, -04:00, -04:00, 282, 283, 647, 755, N, N, H, H, CA, CA, US, US, SN, SN360, CBJNHL, CBJHN, 21, 114, 1, 1, en, NA, NA, NA, 282, NA, NA, NA, https://assets.nhle.com/logos/broadcast/282-light.svg, https://assets.nhle.com/logos/broadcast/283-light.svg, NA, NA, https://assets.nhle.com/logos/broadcast/282-dark.svg, https://assets.nhle.com/logos/broadcast/283-dark.svg, NA, NA, FUT, OK, https://www.ticketmaster.com/event/050064EDB0C08016?brand=bluejackets&artistid=805927&wt.mc_id=NHL_TEAM_CBJ_SCOREBOARD_GM6&utm_source=web_scoreboard_club&utm_medium=NHL.com&utm_campaign=single_game_tickets&utm_content=SCOREBOARD_GM6, https://www.ticketmaster.com/event/050064EDB0C08016?brand=bluejackets&artistid=805927&wt.mc_id=NHL_TEAM_CBJ_SCOREBOARD_GM6&utm_source=web_scoreboard_club&utm_medium=NHL.com&utm_campaign=single_game_tickets&utm_content=SCOREBOARD_GM6, Nationwide Arena, 10, TOR, 2-2-0, https://assets.nhle.com/logos/nhl/svg/TOR_light.svg, Toronto Maple Leafs, Maple Leafs de Toronto, Maple Leafs, Toronto, de Toronto, 29, CBJ, 1-1-0, https://assets.nhle.com/logos/nhl/svg/CBJ_light.svg, Columbus Blue Jackets, Blue Jackets de Columbus, Blue Jackets, Columbus, de Columbus
#> 
# }