PHF Game Raw

phf_game_raw: pull in the raw data for a game_id from the PHF/NWHL API

Usage

phf_game_raw(game_id)

Arguments

game_id

The unique ID code for the game that you are interested in viewing the data for

Value

A list of data frames

Examples

# \donttest{
  try(phf_game_raw(game_id = 612254))
#> [[1]]
#> # A tibble: 126 × 10
#>    ``                      Team  Time  Play  ``    ``    ``    ``    ``    ``   
#>    <chr>                   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Goalie                  Metr… 00:00 Star… NA    NA    NA    NA    NA    NA   
#>  2 Starting Goalie #88 Ka… Star… Star… NA    NA    NA    NA    NA    NA    NA   
#>  3 Goalie                  Toro… 00:00 Star… NA    NA    NA    NA    NA    NA   
#>  4 Starting Goalie #29 El… Star… Star… NA    NA    NA    NA    NA    NA    NA   
#>  5 Faceoff                 Toro… 00:05 #27 … NA    NA    NA    NA    NA    NA   
#>  6 #27 Shiann Darkangelo … #27 … #27 … NA    NA    NA    NA    NA    NA    NA   
#>  7 Giveaway                Toro… 00:40 #27 … NA    NA    NA    NA    NA    NA   
#>  8 #27 Shiann Darkangelo … #27 … #27 … NA    NA    NA    NA    NA    NA    NA   
#>  9 Shot                    Metr… 00:42 #77 … NA    NA    NA    NA    NA    NA   
#> 10 #77 Kennedy Ganser sav… #77 … #77 … NA    NA    NA    NA    NA    NA    NA   
#> # … with 116 more rows
#> 
#> [[2]]
#> # A tibble: 2 × 2
#>   `On Ice` `On Ice`                                                             
#>   <chr>    <chr>                                                                
#> 1 MET      #10 Kendall Cornine #28 Leah Marino #12 Ebba Berglund #21 Amanda Pel…
#> 2 TOR      #10 Lexi Templeman #8 Leah Lum #96 Dominika Lásková #11 Breanne Wils…
#> 
#> [[3]]
#> # A tibble: 2 × 2
#>   `On Ice` `On Ice`                                                             
#>   <chr>    <chr>                                                                
#> 1 TOR      #96 Dominika Lásková #21 Tereza Vanišová #10 Lexi Templeman #8 Leah …
#> 2 MET      #14 Madison Packer #21 Amanda Pelkey #3 Sarah Forster #15 Minttu Tuo…
#> 
#> [[4]]
#> # A tibble: 138 × 10
#>    ``                      Team  Time  Play  ``    ``    ``    ``    ``    ``   
#>    <chr>                   <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Faceoff                 Toro… 00:00 #11 … NA    NA    NA    NA    NA    NA   
#>  2 #11 Breanne Wilson-Ben… #11 … #11 … NA    NA    NA    NA    NA    NA    NA   
#>  3 Shot                    Toro… 00:08 #96 … NA    NA    NA    NA    NA    NA   
#>  4 #96 Dominika Lásková s… #96 … #96 … NA    NA    NA    NA    NA    NA    NA   
#>  5 Faceoff                 Metr… 00:08 #16 … NA    NA    NA    NA    NA    NA   
#>  6 #16 Sarah Bujold faceo… #16 … #16 … NA    NA    NA    NA    NA    NA    NA   
#>  7 Shot                    Toro… 00:58 #96 … NA    NA    NA    NA    NA    NA   
#>  8 #96 Dominika Lásková s… #96 … #96 … NA    NA    NA    NA    NA    NA    NA   
#>  9 Faceoff                 Metr… 01:00 #77 … NA    NA    NA    NA    NA    NA   
#> 10 #77 Kennedy Ganser fac… #77 … #77 … NA    NA    NA    NA    NA    NA    NA   
#> # … with 128 more rows
#> 
#> [[5]]
#> # A tibble: 2 × 2
#>   `On Ice` `On Ice`                                                             
#>   <chr>    <chr>                                                                
#> 1 MET      #94 Fanni Garát-Gasparics #12 Ebba Berglund #8 Kelly Babstock #18 Re…
#> 2 TOR      #41 Brittany Howard #10 Lexi Templeman #71 Saroya Tinker #11 Breanne…
#> 
#> [[6]]
#> # A tibble: 2 × 2
#>   `On Ice` `On Ice`                                                             
#>   <chr>    <chr>                                                                
#> 1 TOR      #41 Brittany Howard #86 Michela Cava #8 Leah Lum #67 Emma Woods #27 …
#> 2 MET      #14 Madison Packer #5 Anna Kilponen #77 Kennedy Ganser #15 Minttu Tu…
#> 
#> [[7]]
#> # A tibble: 82 × 4
#>    ``                                                          Team  Time  Play 
#>    <chr>                                                       <chr> <chr> <chr>
#>  1 Faceoff                                                     Metr… 00:00 #77 …
#>  2 #77 Kennedy Ganser faceoff won against #27 Shiann Darkange… #77 … #77 … NA   
#>  3 Shot                                                        Metr… 00:15 #21 …
#>  4 #21 Amanda Pelkey saved by #29 Elaine Chuli                 #21 … #21 … NA   
#>  5 Faceoff                                                     Toro… 00:15 #27 …
#>  6 #27 Shiann Darkangelo faceoff won against #8 Kelly Babstock #27 … #27 … NA   
#>  7 Shot BLK                                                    Toro… 01:19 #9 K…
#>  8 #9 Kati Tabin blocked by #3 Sarah Forster2 Blk              #9 K… #9 K… NA   
#>  9 Shot                                                        Toro… 01:48 #41 …
#> 10 #41 Brittany Howard saved by #88 Katie Burt                 #41 … #41 … NA   
#> # … with 72 more rows
#> 
#> [[8]]
#> # A tibble: 38 × 4
#>    ``                                                          Team  Time  Play 
#>    <chr>                                                       <chr> <chr> <chr>
#>  1 Faceoff                                                     Toro… 00:00 #27 …
#>  2 #27 Shiann Darkangelo faceoff won against #77 Kennedy Gans… #27 … #27 … NA   
#>  3 Shot                                                        Toro… 00:40 #27 …
#>  4 #27 Shiann Darkangelo saved by #88 Katie Burt               #27 … #27 … NA   
#>  5 Shot                                                        Metr… 01:19 #16 …
#>  6 #16 Sarah Bujold saved by #29 Elaine Chuli                  #16 … #16 … NA   
#>  7 Shot BLK                                                    Metr… 01:25 #5 A…
#>  8 #5 Anna Kilponen blocked by #21 Tereza Vanišová1 Blk        #5 A… #5 A… NA   
#>  9 Shot                                                        Metr… 02:11 #12 …
#> 10 #12 Ebba Berglund saved by #29 Elaine Chuli                 #12 … #12 … NA   
#> # … with 28 more rows
#> 
#> [[9]]
#> # A tibble: 1 × 4
#>   ``                Team              Time              Play             
#>   <chr>             <chr>             <chr>             <chr>            
#> 1 No plays recorded No plays recorded No plays recorded No plays recorded
#> 
#> [[10]]
#> # A tibble: 5 × 3
#>   ``    Team                  Play                                              
#>   <chr> <chr>                 <chr>                                             
#> 1 Miss  Toronto Six           #96 Dominika Lásková missed attempt against #88 K…
#> 2 Miss  Metropolitan Riveters #94 Fanni Garát-Gasparics missed attempt against …
#> 3 Miss  Toronto Six           #41 Brittany Howard missed attempt against #88 Ka…
#> 4 Goal  Metropolitan Riveters #15 Minttu Tuominen scores against #29 Elaine Chu…
#> 5 Miss  Toronto Six           #11 Breanne Wilson-Bennett missed attempt against…
#> 
#> [[11]]
#> # A tibble: 5 × 3
#>   ``    Team                  Play                                              
#>   <chr> <chr>                 <chr>                                             
#> 1 Miss  Toronto Six           #96 Dominika Lásková missed attempt against #88 K…
#> 2 Miss  Metropolitan Riveters #94 Fanni Garát-Gasparics missed attempt against …
#> 3 Miss  Toronto Six           #41 Brittany Howard missed attempt against #88 Ka…
#> 4 Goal  Metropolitan Riveters #15 Minttu Tuominen scores against #29 Elaine Chu…
#> 5 Miss  Toronto Six           #11 Breanne Wilson-Bennett missed attempt against…
#> 
#> [[12]]
#> # A tibble: 2 × 7
#>   Scoring `1st` `2nd` `3rd`    OT SO            T
#>   <chr>   <int> <int> <int> <int> <chr>     <int>
#> 1 MET         1     1     0     0 1 (1 - 1)     3
#> 2 TOR         1     1     0     0 0 (0 - 3)     2
#> 
#> [[13]]
#> # A tibble: 2 × 6
#>   Shots `1st` `2nd` `3rd`    OT     T
#>   <chr> <int> <int> <int> <int> <int>
#> 1 MET      14    10    12     2    39
#> 2 TOR       8    14     5     5    32
#> 
#> [[14]]
#> # A tibble: 7 × 3
#>   `Team Stats`           MET   TOR  
#>   <chr>                  <chr> <chr>
#> 1 Power Plays            0 / 5 2 / 4
#> 2 Power Play %           .000  .500 
#> 3 Penalty Minutes        6     10   
#> 4 Faceoff %              .516  .484 
#> 5 Blocked Opponent Shots 13    11   
#> 6 Takeaways              9     3    
#> 7 Giveaways              3     9    
#> 
# }